LayoutTests/fast/forms/file/webkitdirectory-open-panel.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/ui-helper.js"></script>
</head>
<body>
<input type="file" onchange="changed(event)" webkitdirectory></input>
<script>
description("Tests choosing a folder via the file picker when using webkitdirectory.");
jsTestIsAsync = true;

function runTest()
{
    testRunner.setOpenPanelFiles(['resources/testFiles']);

    inputElement = document.getElementsByTagName('input')[0];
    shouldBeTrue("inputElement.webkitdirectory");
    UIHelper.activateAt(inputElement.offsetLeft + inputElement.offsetWidth / 2,
        inputElement.offsetTop + inputElement.offsetHeight / 2);
}

function changed(event)
{
    fileList = event.target.files;
    shouldBe("fileList.length", "5");
    shouldBeEqualToString("fileList[0].name", "file1.txt");
    shouldBeEqualToString("fileList[0].webkitRelativePath", "testFiles/file1.txt");
    shouldBeEqualToString("fileList[1].name", "file2.txt");
    shouldBeEqualToString("fileList[1].webkitRelativePath", "testFiles/file2.txt");
    shouldBeEqualToString("fileList[2].name", "file3.txt");
    shouldBeEqualToString("fileList[2].webkitRelativePath", "testFiles/subfolder1/file3.txt");
    shouldBeEqualToString("fileList[3].name", "file4.txt");
    shouldBeEqualToString("fileList[3].webkitRelativePath", "testFiles/subfolder2/file4.txt");
    shouldBeEqualToString("fileList[4].name", "file5.txt");
    shouldBeEqualToString("fileList[4].webkitRelativePath", "testFiles/subfolder2/subfolder2a/file5.txt");

    finishJSTest();
}

runTest();

</script>
</body>
</html>
